Moved models urls to utils folder
