+++ src/app/journey/03-advanced-techniques/02-imported-models/02-imported-models.ts
@@ -14,7 +14,7 @@ import { Timer } from 'three/addons/misc/Timer.js';
 import { ActionButton, Description, Exercise } from '#/app/decorators/exercise';
 import RenderView from '#/app/layout/render-view';
 import { AssetLoader } from '#/app/services/assets-loader';
-import { DUCK_URL, FOX_URL, HELMET_URL } from '#/app/utils/tests/models-urls';
+import { DUCK_URL, FOX_URL, HELMET_URL } from '#/app/utils/models-urls';
 import { disposeMesh, disposeObjects } from '#/app/utils/three-utils';
 import DUCK from './icons/duck.svg?raw';
 import FOX from './icons/fox.svg?raw';
