+++ src/app/journey/03-advanced-techniques/04-custom-models.ts
@@ -14,7 +14,7 @@ import { Description, Exercise } from "#/app/decorators/exercise";
 import OrbitControlledExercise from "#/app/journey/exercises/orbit-controlled-exercise";
 import RenderView from "#/app/layout/render-view";
 import { AssetLoader } from "#/app/services/assets-loader";
-import { HAMBURGER_URL } from '#/app/utils/tests/models-urls';
+import { HAMBURGER_URL } from '#/app/utils/models-urls';
 import { disposeMesh } from "#/app/utils/three-utils";
 
 @Exercise("Hamburger")
