+++ src/app/journey/03-advanced-techniques/07-realistic-render/index.ts
@@ -18,7 +18,7 @@ import OrbitControlledExercise from "#/app/journey/exercises/orbit-controlled-ex
 import RenderView from "#/app/layout/render-view";
 import { AssetLoader } from "#/app/services/assets-loader";
 import { ExtraConfig } from '#/app/types/exercise';
-import { HAMBURGER_URL } from '#/app/utils/tests/models-urls';
+import { HAMBURGER_URL } from '#/app/utils/models-urls';
 import { loadTextureMaps, TextureDict, TextureMaps } from "#/app/utils/textures";
 import { disposeMesh, disposeObjects } from "#/app/utils/three-utils";
 import { SCENE_CONTROLLERS, RENDERER_CONTROLLERS, LIGHT_CONTROLLERS } from "./controllers";
