+++ src/app/journey/03-advanced-techniques/08-animation-mixer/index.ts
@@ -21,7 +21,7 @@ import OrbitControlledExercise from "#/app/journey/exercises/orbit-controlled-ex
 import RenderView from "#/app/layout/render-view";
 import { AssetLoader } from "#/app/services/assets-loader";
 import { ExtraConfig } from "#/app/types/exercise";
-import { FOX_URL } from "#/app/utils/tests/models-urls";
+import { FOX_URL } from "#/app/utils/models-urls";
 import { disposeMesh } from "#/app/utils/three-utils";
 import IDLE from './icons/idle.svg?raw';
 import RUN from './icons/run.svg?raw';
