+++ src/app/journey/05-extra/01-post-processing/index.ts
@@ -21,7 +21,7 @@ import { Description, Exercise } from "#/app/decorators/exercise";
 import { EnvironmentMap } from "#/app/journey/common/environment-map";
 import RenderView from "#/app/layout/render-view";
 import { AssetLoader } from "#/app/services/assets-loader";
-import { DAMAGED_HELMET_URL } from "#/app/utils/tests/models-urls";
+import { DAMAGED_HELMET_URL } from "#/app/utils/models-urls";
 import displacementFrag from "./shaders/displacement.frag";
 import displacementVert from "./shaders/displacement.vert";
 import tintFrag from "./shaders/tint.frag";
